Thread Engine
I am starting on changing the linked web views to a single web view
based on a template. It does not seem that we can mix and match WinJS
and C#, but we "should" be able to invoke C# from javascript functions.
Right now I am working on adding post previews to the reply page using
this engine. It is not working yet, but once it works there we can
change the main thread view page to use it. This will increase
performance and lower memory usage on that view.