Change rapsheet/user post history list views to item controllers
This change is temporary. Right now the post history/rapsheet pages
don't go anywhere, and are only used to SHOW you their respected
information. If we use a list view, and Microsoft sees a list view that
does not go anywhere when tapped, they will fail the app.

When we come up with some other view to go to when tapped, then we can
switch it back. We can also restyle it anyway, because it looks pretty
poor.