Revamp Reply/Edit Views
I've gotten rid of the previous edits to reply view and instead rewrote
the XAML. All buttons are now moved to the left navigation grid. The
reply box is still on the top left, but the Smilies/BBCode/Preview/Last
Posts sections are all on the bottom left part of the grid. This should
simplify reply/edit operations and give us better snap support for free.