Capture SA Forum Links and redirect back to app.
On the thread webview, links are now being captured in javascript. If
the link is for something inside the forums, it needs to remain in the
app rather than open IE.

If you link to a thread, user profile, post history or rap sheet, it
should redirect the user to the functions of the app. It should also
have basic error handling in case the URL is malformed.

All other links should be handled as they were before; by opening IE.